+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## v0.158.0 (2026-07-15)
+
+### Feat
+
+- **devices-manager**: named polling groups for granular intervals (#456)
+
+## v0.157.0 (2026-07-15)
+
+### Feat
+
+- **devices-manager**: add read_many streaming contract (#455)
+
 ## v0.156.0 (2026-07-15)
 
 ### Feat